requirements-onnx.txt
# Optimized Backend Requirements with ONNX (Replaces PyTorch YOLO)
# Total deployment size: ~1.5 GB (vs 8.2 GB with ultralytics)
# Full CT scan AI functionality preserved!

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0

# ONNX Runtime (Lightweight inference - replaces ultralytics/PyTorch)
onnxruntime>=1.16.0

# Image Processing (headless version for server deployment)
opencv-python-headless>=4.8.0
pillow>=10.2.0
numpy>=1.24.0

# Medical Image Support (DICOM)
pydicom>=2.4.0

# Database
psycopg2-binary>=2.9.9

# Configuration & Environment
python-dotenv>=1.0.0

# Async File Operations
aiofiles>=23.2.1
